fixed multiple errors. Need to get login working
