+++ codesnippets_project/code_snippets/urls.py
@@ -3,6 +3,4 @@ from . import views
 
 urlpatterns = [
     path('', views.snippets, name='snippets'),
-    path('<int:pk>/', views.userprofile, name='userprofile'),
-    
 ]
