--- codesnippets_project/templates/login_error.html
@@ -1,1 +1,0 @@
-{% extends "base.html" %}
