+++ codesnippets_project/templates/login_user.html
@@ -19,7 +19,7 @@
         <input type="password" name="password">
     </p>
     <div>
-        <button type="submit">Login</button>
+        <input type="submit" value="login" />
     </div>
 </form>
 <div>
