--- codesnippets_project/templates/logout.html
@@ -1,5 +1,0 @@
-{% extends "base.html"%} 
-
-<h1>YOU ARE LOGGED OUT</h1>
-
-{% endblock %}
