+++ codesnippets_project/templates/logout_user.html
@@ -1,0 +1,9 @@
+{% extends "base.html"%} 
+
+{% block content %}
+
+<h1>YOU ARE LOGGED OUT</h1>
+<br>
+<p><a class="link" href="{% url 'snippets' %}"> Back to Snippets</a></p><br>
+
+{% endblock %}
