+++ codesnippets_project/templates/snippets.html
@@ -9,10 +9,10 @@
 {% block nav %}
 {% if logged_in %}
 <span class="link"><a href="{% url 'userprofile' pk=user.pk %}">Profile</a></span>
-<span class="link"><a href="{% url 'snippets' %}">Logout</a></span>
+<span class="link"><a href="{% url 'logout_user' %}">Logout</a></span>
 {% endif %}
 {% if not logged_in %}
-<span class="link"><a href="{% url 'login' %}">Login</a></span>
+<span class="link"><a href="{% url 'login_user' %}">Login</a></span>
 <span class="link"><a href="{% url 'register' %}">Register</a></span>
 {% endif %}
 {% endblock %}
