+++ codesnippets_project/templates/userprofile.html
@@ -3,7 +3,7 @@
 {% block content %}
 <p><a class="link" href="{% url 'snippets' %}"> Back to Snippets</a></p><br>
 
-<h1> User snippets </h1>
+<h1> Hello {{ user }}, welcome to your Snippets </h1>
