+++ codesnippets_project/users/urls.py
@@ -4,6 +4,7 @@ from . import views
 
 urlpatterns = [
         path('register/', views.register, name='register'),
-        path('login/', views.login, name='login'),
-        path('logout/', views.logout, name='logout'),
+        path('login/', views.login_user, name='login_user'),
+        path('logout/', views.logout_user, name='logout_user'),
+        path('<int:pk>/', views.userprofile, name='userprofile'),
 ]
